refactor: story code
